GIR writer: Fix generation of implemented interfaces
The list of implemented interfaces is expected as

    <implements name="Interface1" />
    <implements name="Interface2" />
    ...

by gobject-introspection library (as of 0.6.3), but was written as

    <implements>
        <implements name="Interface1" />
        <implements name="Interface2" />
        ...
    </implements>

by Vala.GIRWriter. Note, that vapigen expects the same format as
gobject-introspection, so it was not able to read vala-written .girs
before the change and is able to read them properly now.

Fixes bug 584576.

Signed-off-by: Jan Hudec <bulb@ucw.cz>