valac: Output make-style dependency file
Add a --deps= option to the compiler to write out a make-style
dependency file.  The name of the target used is the name of the
dependency file itself.

This lets the dependency file serve as a stamp for the C file (which may
or may not be touched depending if it was changed).  The dependency
output is always touched.