SourceFile: change the definition of 'used'
Consider a source file as used only if actual C code (or declarations)
have been generated from something in that source file.

This greatly reduces the number of dependencies listed for compilations
that touch a large number of symbols in trivial ways without really
making use of them (for example, when listing a very large number of
unused fast-vapi files, as parvala does).

Fixes bug 630144.