codegen: check return value before ref_sink the value
When a function is expected to give back a floating reference, the
caller is trying to ref_sink the reference after calling this function.
But when the function fails and returns NULL, the ref_sink function is
still called. This can lead to a runtime warning as the ref_sink
function is not expecting a NULL value. Additionally remove all
ref_sink_function_void, as this is only for C convenience.