gio-2.0: fix Settings.bind_with_mapping
This method is skipped in the GIR because of the shared user_data
parameter.  Vala doesn't support this, either, so to get around it
we have to set has_target = false on the delegates and pass the void*
manually.  This also changes the names of the relevant delegates so
there is no collision if/when the issue is resolved in GIO.