doclets/gtkdoc: Handle ‘Deprecated’ attributes in the gtk-doc doclet
On visiting a ‘Deprecated’ attribute, emit a ‘Deprecated’ header
in the gtk-doc comment. This includes logic to warn the user if
the ‘since’ or ‘replacement’ arguments are missing from the
attribute, and suitable boilerplate if they are.

This means that deprecated Vala symbols are now correctly listed
in the gtk-doc deprecation index.