linux: make backtrace_symbols return unowned
You are supposed to free() the result of backtrace_symbols, but not
each element in the array.  The current binding will free each element
which will likely result in a crash.  Since we don't currently support
owned arrays of unowned values (bug #571486), making the whole thing
unowned and leaking (unless you call free manually) is the best we can
do right now.