gen-introspect: pretend to be lint
Some system headers (notably <stdarg.h> on FreeBSD) get upset when they
see an unexpected compiler.  We go out of our way to pass -U__GNUC__ to
cpp from the scanner code during gen-introspect and this is causing
problems.

As a workaround, we can pretend to be lint, which fixes the issue at
least for FreeBSD.

https://bugzilla.gnome.org/show_bug.cgi?id=722954