girparser: grok "nullable"
Understand the new "nullable" attribute in .gir files.

Presently, this is mostly an alias for allow-none='', but it will also
allows a new feature: we can explicitly mark out parameters as having a
nullable type (as a distinct concept from accepting 'NULL' as a
parameter to the C function call in order to ignore the result).

.gir may eventually want to remove allow-none='' some day, so make sure
we also accept nullable='' in all places that we accepted allow-none=''
to mean 'can be NULL'.

https://bugzilla.gnome.org/show_bug.cgi?id=660879