girparser: don't deprecate children of deprecated containers
GObject Introspection doesn't support marking entire types as
deprecated, so we end up with a lot of places where the type isn't
deprecated, but every method, property, field, signal, etc. in it is,
which causes a lot of spew during compilation.  This patch allows us
to easily deprecate a type in metadata and any extra deprecation
information on child symbols will be omitted.