libsoup-2.4: deprecate Buffer.new and MessageBody.append
These functions require the user to correctly specify a Soup.MemoryUse
argument, but that's easy to get wrong without understanding how the
memory is managed.  Alternatives for these functions which will always
transfer ownership of the memory have been in place since 2.32, so
from now on we will emit a warning when someone tries to use the old
versions.