gstreamer-rtsp-server-1.0: update bindings
The previous bindings only worked for an unstable pre-release version
of gstreamer-rtsp-server-1.0.  The new (stable) version is not
backwards compatible, so neither can the Vala bindings be.  Since we
had to break backwards compatibilitiy, why not go big?  I've taken the
opportunity to move the bindings into the Gst.RTSPServer namespace to
better match the other gstreamer bindings, as well as other assorted
less-disruptive clean-ups.

Fixes bug 723910.