libnl-3.0.vapi: extend VAPI
The patch extends VAPI of libnl-3.0 with cache
manager (nl_cache_mngr_*), additional methods of LinkCache and a helper
CachedLink type. The CachedLink type is a wrapper for the Link object,
however since the object is obtained from cache, rtnl_link_put() must be
used a free function.

Signed-off-by: Maciej Borzecki <maciek.borzecki@gmail.com>
Signed-off-by: Maciej Borzecki <maciej.borzecki@open-rnd.pl>

Fixes bug 763303