vala: Don't force array-elements to be owned, unowned ones are supported
Passing an (unowned string) array literal to a function causes the array
elements to be freed afterwards, even though they should not be. This
causes the array elements to be double freed. While assigning the literal
to an intermediate variable and passing that to the function masks this
error of the ArrayCreationExpression.

https://bugzilla.gnome.org/show_bug.cgi?id=761307