girwriter: Add fullname GIR attribute, and set fullname of GObject.Type
In gir files generated by Vala, members of type GType is referenced
with name "GObject.Type", but GObject-introspection expect it to be
"GType".

While there was no way to override the fully qualified GIR name of a
member, this patch adds [GIR (fullname = "GType")] to fix the mismatch
for GObject.Type.

https://bugzilla.gnome.org/show_bug.cgi?id=775591