codegen: Compensate for the lack of g_task_get_completed on glib < 2.44
If a target glib between 2.36 and 2.44 is used, store a boolean var
in the async task data, which is initialized to FALSE and set to
TRUE in the async task callback, so async-forced-to-sync like the
async generator example can block until the task is complete.

There is one special case, if the async task receives no callback
to execute, the boolean flag is set to TRUE right away, as it
will be "finished" by the first time it's checked.

https://bugzilla.gnome.org/show_bug.cgi?id=763345