codegen: Make the task_complete flag for < 2.44 more similar to >= 2.44
According to the g_task_get_completed() docs (which we rely on for
glib >= 2.44 targets): "This changes from FALSE to TRUE after the task's
callback is invoked, and will return FALSE from inside the callback".

So to make the code paths most similar to >= 2.44 (when
g_task_get_completed is available), wrap the GAsyncReadyCallback with
one of our own, that just invokes the nested callback (if any) and turns
on the flag.

Also remove the code turning on the flag on finish(), it's superfluous
now and there are no guarantees that it will be invoked.

https://bugzilla.gnome.org/show_bug.cgi?id=783543