codegen: Free generic elements of glib collections
It needs to be possible to use parameters or fields/properties which hold
dup/free functions for a generic type in scope.

This required to make the destroy_func being a parameter with the benefit
of being able to use g_*_free_all directly and adding a _g_node_free_all
wrapper with a compatible signature.

https://bugzilla.gnome.org/show_bug.cgi?id=694765