glib-2.0: fix g_stat () binding.
The data type should be GStatBuf, not struct stat.

The documentation of g_stat () states:

In Microsoft's compiler, by default struct stat means one with
64-bit time fields while in MinGW struct stat is the legacy one
with 32-bit fields.

To hopefully clear up this messs, the gstdio.h header defines a
type GStatBuf which is the appropriate struct type depending
on the platform and/or compiler being used. On POSIX it is just
struct stat, but note that even on POSIX platforms, stat() might
be a macro.