codegen: Don't set implemenation of interface property to its own
A base-class is allowed to provide interface implementations for methods
and properties. If those exist without an explicit implementation in the
sub-class we then we end up finding the interface property itself. Using
that is obviously wrong and causes a cyclic call stack.

Fixes https://gitlab.gnome.org/GNOME/vala/issues/891