Add 'opaque' compact classes
This change now allows to completely hide the implementation of a
compact class, which is especially useful for libraries when maintaining
a stable abi. An 'opaque' compact class exposes no struct definition in
the generated c header, only a typedef is provided.

This lifts the previous restriction of compact classes not having
private fields, with the following caveats:
- A compact class may not mix private and non-private fields.
- A compact class may not have both private fields and abstract/virtual
  methods/properties.

In both cases, an apropriate error is raised during compilation.