Add 'opaque' compact classes
This change intruduces a new attribute switch [Compact (opaque=true)]
which allows to completely hide the implementation of a compact class.
This is especially useful for libraries when maintaining a stable abi.

An 'opaque' compact class exposes no struct definition in the generated
c header, only a typedef is provided. As such, certain requirements
apply for members of such classes:
- Access to instance fields must be either private or internal.
- No abstract/virtual methods or properties are allowed.