Fix #line directives for external packages
Currently, code that is declared inline in vapis emits only the files
basename. This change assigns the vapis path to relative_filename.

Note: If the vapidir used to resolve the package was set to an absolute
path, relative_filename will also be an absolute path. This might be
preferable for system packages but undesired for project internal vapi
directories. Also note that meson requires all vapidirs to be absolute.