Fix #line directive numbering
Currently, various c statements are generated that do not point to the
actual vala source code. In such cases, vala generates a line referring
back to the c file. While this is mildly annoying during debugging, it
also leads to incorrect coverage reports.

This change attempts to always emit line directives referring to the
actual vala source line. At the end of function blocks, a line directive
is then generated, leading back to the c file.