vapi/glib-2.0.vapi: move cname to CCode attribute for canonicalize
Currently GLib.canonicalize has the cname attribute in Version.

However the cname attribute applies only to CCode so the compiler thinks the corresponding cname
is g_filename_canonicalize, which isn't implemented in glib under this name, leading to undefined
reference to g_filename_canonicalize.

To fix this, move the cname pair to a CCode attribute for canonicalize.

(No gitlab issue as trivial)