vala: Add multi-line support to Vala.Report
To fix multi-line errors, the "^~~~"-underline appears now under all lines
and columns between the begin and end of the source reference.

Also a line number with a "|" at the left side is added for better finding
the source of the error. This helps also one-line errors.

Fixes https://gitlab.gnome.org/GNOME/vala/issues/764