valaccompiler.vala: Check for MSVC-like compiler
If we are using an enviroment where it is set up to be like a Visual Studio
cmd.exe prompt, and if CC is not set, we assume that we are using Visual Studio

Thus, we need to call pkg-config with the --msvc-syntax flag.  This also apples
if we choose to use clang-cl, so that we can use the correct C compiler/linker
flags when we compile the generated C-code.