Add no_generics_args attribute
The no_generics_args attribute can be added to anything that can have
generics, for example classes, interfaces and methods. It makes the C
code generation not emit parameters, properties fields, etc. for the
storage or passing over of type parameters. This means 3 things: the
type id, copy function and free function.

The attribute works in .vapi files as well as full implementations in
.vala files.

But these values still need to be taken from somewhere to for example
copy an object with the type of the type parameter. For this reason it
is required that there is a constrain for the generic parameter. The
constrain needs to be an "object" type. So a type derived from
GTypeInstance. This can be standard vala classes with or without
GLib.Object as a base or Interfaces. It is because these types provide
ref- and unref-functions. These together with the type id are used as
constants instead of the variable type passed through parameters.

The attribute applied on a type makes all its type parameters
not emit no generic arguments.