flatpak: Don't clean /share/aclocal
Without the autoconf macros, we have no chance of building autotools
projects that depend on vala. Missing macros results in exceptionally
confusing error messages. For example, this error when building libgee:

./configure: line 13551: syntax error near unexpected token `fi'
./configure: line 13551: `fi'

This is going to affect almost all Vala applications, except for the few
that do not use libgee, because it's no longer provided by the runtime
and it uses Autotools.

https://github.com/flathub/org.freedesktop.Sdk.Extension.vala/pull/91